requirements.txt
google-cloud-bigquery[pandas]==3.38.0
pandas==2.2.3
python-dotenv
pandas
pyarrow
db-dtypes